+++ src/QmlControls/RCChannelMonitorController.cc
@@ -19,7 +19,7 @@ RCChannelMonitorController::RCChannelMonitorController(QObject *parent)
 {
     // qCDebug(RCChannelMonitorControllerLog) << Q_FUNC_INFO << this;
 
-    (void) connect(_vehicle, &Vehicle::rcChannelsChanged, this, &RCChannelMonitorController::_rcChannelsChanged);
+    (void) connect(_vehicle, &Vehicle::rcChannelsChanged, this, &RCChannelMonitorController::channelValuesChanged);
 }
 
 RCChannelMonitorController::~RCChannelMonitorController()
@@ -27,7 +27,7 @@ RCChannelMonitorController::~RCChannelMonitorController()
     // qCDebug(RCChannelMonitorControllerLog) << Q_FUNC_INFO << this;
 }
 
-void RCChannelMonitorController::_rcChannelsChanged(int channelCount, int pwmValues[QGCMAVLink::maxRcChannels])
+void RCChannelMonitorController::channelValuesChanged(int channelCount, int pwmValues[QGCMAVLink::maxRcChannels])
 {
     for (int channel = 0; channel < channelCount; channel++) {
         const int channelValue = pwmValues[channel];
@@ -38,7 +38,7 @@ void RCChannelMonitorController::_rcChannelsChanged(int channelCount, int pwmVal
         }
 
         if (channelValue != -1) {
-            emit channelRCValueChanged(channel, channelValue);
+            emit channelValueChanged(channel, channelValue);
         }
     }
 }
